Curved Text Character Angle
I have make changes which will solve unknown reason of divide by 4 (it must be 2).
You must set 'curAngle' value as angle of character instead of 'curAngleRotation'. So you don't need to set curAngleRotation.
To test and see difference of this change, set radius to '200' and spacing to '575' with both files.